Make `close_file` and `add_chunk` more rigorous
Not a good idea IMO to be lenient with passed values -- a stray `None`
that is "falsy" will cause e.g. `close_file` to be a no-op, and noone
will be wiser, since it's just silent in that way. We want exceptions in
the logs because we aren't planning to pass `None` to `close_file` etc,
normally -- I checked the call sites and there's no reason a `None` is
ever to be expected by either of the procedures. So if `None` ever ends
up there for whatever reason, we want to rightfully be alarmed (through
the logs).