Declare variable for rare code path
In a rare case `size_before_merge` ends up not declared (since an
error may be raised before the initial declaring assignment statmeent)
and then a subsequent attempt to use it will raise another error. This
commit adds a default value so that it in the worst case the null value
at least prevents errors from being raised.

MyPy rightfully complained too. Not that it's part of the project, but I
ran it to help me find potential avenues for troubleshooting the
corruption issue, and this was one of the things it notified me about.