Make an `except` block less "chatty"
No need, IMO, to log an error for an exception when caller is expected
to handle the problem, especially since the error is raised anyway after
the resumable object cleans up in the `except` block. This also makes it
possible to not have to use `e` but just `raise` which re-raises the
error.