Make two `except` blocks more "idiomatic"
Calls to logger methods (`debug`, `error` etc) don't take _errors_ --
although the value is "converted" implicitly with `str`, but in case
with an error that has a stack trace, `exception` method is more apt
since it will print everything itself anyway, with less code explicitly
there. We use the opportunity to provide a meaningful error message
instead -- that pertains to what actually was attempted that raised the
error.